+++ R/Zdeprecated.R
@@ -4,7 +4,7 @@
 if (!isGeneric("convexhull")) {setGeneric("convexhull", function(x, ...) standardGeneric("convexhull"))}
 setMethod("convexhull", signature(x="SpatVector"), 
 	function(x) {
-		error("convexhull", "terra::convexhull has been removed. Use 'convhull' instead")
+		error("convexhull", "terra::convexhull has been removed. Use 'convHull' instead")
 	}
 )